requirements.txt
streamlit
google-genai
google-cloud-storage
python-dotenv
requests
beautifulsoup4
pypdf